+++ src/craft/craft.c
@@ -2896,7 +2896,7 @@ ACMD(do_harvest)
   definition.progress_model = PRIMARY_ACTIVITY_PROGRESS_PROGRESSIVE;
   definition.progress_owner = PRIMARY_ACTIVITY_PROGRESS_CHARACTER;
   definition.total_steps = NODE_HARVEST_STEPS;
-  definition.step_interval = PULSE_VIOLENCE;
+  definition.step_interval = (long)PULSE_VIOLENCE;
   definition.combat_actions_required = ACTION_STANDARD | ACTION_MOVE;
   definition.movement_response = PRIMARY_ACTIVITY_RESPONSE_CANCEL;
   definition.damage_response = PRIMARY_ACTIVITY_RESPONSE_CANCEL;
